+++ src/mra_net.h
@@ -61,7 +61,7 @@ typedef struct _mra_serv_conn {
     void (*callback_login)(gpointer, uint32_t, gchar*);                                                 // auth
     void (*callback_logout)(gpointer, gchar*);                                                          // auth
     void (*callback_user_info)(gpointer, mra_user_info*);                                               // user info
-    void (*callback_contact_list)(gpointer, uint32_t, uint32_t, mra_group*, uint32_t, mra_contact*);    // contact list
+    void (*callback_contact_list)(gpointer, uint32_t, size_t, mra_group*, size_t, mra_contact*);    // contact list
     void (*callback_user_status)(gpointer, char*, uint32_t);                                            // user status
     void (*callback_auth_request)(gpointer, char*, char*);                                              // auth request
     void (*callback_typing_notify)(gpointer, char*);                                                    // typing notify
@@ -78,7 +78,7 @@ char *mra_net_mklps(const char *);
 char *mra_net_mksz(char *);
 
 void mra_net_fill_cs_header(mrim_packet_header_t *, uint32_t, uint32_t, uint32_t);
-void mra_net_send(gpointer, gpointer, uint32_t);
+void mra_net_send(gpointer, gpointer, size_t);
 gboolean mra_net_send_flush(gpointer);
 
 gboolean mra_net_ping_timeout_cb(mra_serv_conn *);
@@ -98,12 +98,12 @@ gboolean mra_net_send_anketa_info(mra_serv_conn *, const char *);
 void mra_net_read_cb(gpointer, gint, PurpleInputCondition);
 gboolean mra_net_read_proceed(gpointer);
 
-void mra_net_read_hello(gpointer, char *, uint32_t);
-void mra_net_read_login_successful(gpointer, char *, uint32_t);
-void mra_net_read_login_failed(gpointer, char *, uint32_t);
-void mra_net_read_logout(gpointer, char *, uint32_t);
-void mra_net_read_user_info(gpointer, char *, uint32_t);
-void mra_net_read_contact_list(gpointer, char *, uint32_t);
+void mra_net_read_hello(gpointer, char *, size_t);
+void mra_net_read_login_successful(gpointer, char *, size_t);
+void mra_net_read_login_failed(gpointer, char *, size_t);
+void mra_net_read_logout(gpointer, char *, size_t);
+void mra_net_read_user_info(gpointer, char *, size_t);
+void mra_net_read_contact_list(gpointer, char *, size_t);
 void mra_net_read_user_status(gpointer, char *, uint32_t);
 void mra_net_read_message(gpointer, char *, uint32_t);
 void mra_net_read_message_status(gpointer, char *, uint32_t);
